moved duplicate functions to the auxiliary_methods file
