+++ backend/src/core/ocim/concurrent_cut.rs
@@ -1,6 +1,8 @@
 use rustc_hash::FxHashSet;
 
-use crate::core::ocim::auxiliary_methods::{get_projected_end, get_projected_start};
+use crate::core::ocim::auxiliary_methods::{
+    get_projected_end, get_projected_start, partitions_cover_alphabet,
+};
 use crate::core::ocim::common_data::{GlobalData, LocalData};
 
 /// Rust port of the Python `is_concurrent_cut_valid` helper.
@@ -109,12 +111,6 @@ pub fn is_concurrent_cut_valid(
     }
 
     true
-}
-
-fn partitions_cover_alphabet(partitions: &[Vec<String>], alphabet: &[String]) -> bool {
-    let part_set: FxHashSet<_> = partitions.iter().flatten().cloned().collect();
-    let alphabet_set: FxHashSet<_> = alphabet.iter().cloned().collect();
-    part_set == alphabet_set
 }
 
 #[cfg(test)]
