+++ backend/src/core/ocim/concurrent_cut_detection.rs
@@ -1,6 +1,8 @@
 use rustc_hash::{FxHashMap, FxHashSet};
 
-use crate::core::ocim::auxiliary_methods::{get_projected_end, get_projected_start};
+use crate::core::ocim::auxiliary_methods::{
+    components_from_unionfind, get_projected_end, get_projected_start,
+};
 use crate::core::ocim::common_data::{GlobalData, LocalData};
 use crate::core::ocim::concurrent_cut::is_concurrent_cut_valid;
 use crate::models::ocpt::OCPTOperatorType;
@@ -318,15 +320,6 @@ pub fn find_cut_concurrent(
     }
 
     None
-}
-
-fn components_from_unionfind(uf: &petgraph::unionfind::UnionFind<usize>, items: &[String]) -> Vec<Vec<String>> {
-    let mut groups: FxHashMap<usize, Vec<String>> = FxHashMap::default();
-    for (idx, item) in items.iter().enumerate() {
-        let root = uf.find(idx);
-        groups.entry(root).or_default().push(item.clone());
-    }
-    groups.into_values().collect()
 }
 
 #[cfg(test)]
