+++ backend/src/core/ocim/exclusive_cut.rs
@@ -2,6 +2,7 @@ use rustc_hash::FxHashSet;
 
 use crate::core::ocim::auxiliary_methods::{
     get_divergent_types, get_non_divergent_types, get_projected_end, get_projected_start,
+    partitions_cover_alphabet,
 };
 use crate::core::ocim::common_data::{GlobalData, LocalData};
 
@@ -143,12 +144,6 @@ pub fn is_exclusive_cut_valid(
     }
 
     true
-}
-
-fn partitions_cover_alphabet(partitions: &[Vec<String>], alphabet: &[String]) -> bool {
-    let part_set: FxHashSet<_> = partitions.iter().flatten().cloned().collect();
-    let alphabet_set: FxHashSet<_> = alphabet.iter().cloned().collect();
-    part_set == alphabet_set
 }
 
 #[cfg(test)]
